refactor(design-mode): remove write-back — the agent is the only writer
Product decision: design mode writing source files directly will inevitably
race the agent's own edits; rather than arbitrating that conflict (write
queues, verify loops, rollback semantics), design mode is now write-free.
Its job is pointing and intent capture: select an element, describe the
change in the in-page bubble, and the annotation (note + cropped screenshot
+ source/component context) lands in the composer for the agent to execute.

Removed:
- the deterministic write-back engine (tailwind-map, write-plan,
  source-locator, patch, file-write-queue, verify-loop, sourcemap) and its
  tests; @babel/parser + magic-string dependencies dropped
- DesignDrawer (style sliders / Apply / undo) and the bubble's "Aa" opener;
  a headless DesignAnnotateBridge now runs the annotate pipeline
- apply/undo/rollback/preview IPC surface; DesignApplyInput/Result types

Kept: inspector inject/re-inject, fiber + data-attr source tiers, component
chain, annotate bubble, capability probes, localhost gate, suspend pin,
host-reload native-view cleanup.

The full write-back implementation (with its 4-perspective review fixes and
tests) remains in git history (bb12123^..1c49c77) if it is ever revived;
the plan doc records the decision and pointer.

verify:design-writeback renamed to verify:design-mode (annotate logic +
write-free guarantee + wiring); verify:design-e2e now exercises the live
annotate loop (inject → select → bubble → event with context + geometry).
Full test suite passes.
