+++ package.json
@@ -14,9 +14,9 @@
     "dev:electron": "npm run transpile:electron && node scripts/prepare-electron-native-deps.mjs && node scripts/launch-electron.mjs",
     "transpile:electron": "tsc -p src/electron/tsconfig.json",
     "build": "npm run transpile:electron && vite build",
-    "test": "npm run validate:terminal && npm run check:icons && npm run verify:review-diff-panel && npm run verify:workstream-stages && npm run verify:workstream-preamble-order && npm run verify:picker-menu-side && npm run verify:environment-hub-anchor && npm run verify:git-branch-pill && npm run verify:side-chat-drop && npm run verify:opencode-sdk-adapter && npm run verify:pi-sdk-adapter && npm run verify:codex-compaction && npm run verify:spreadsheet-preview && npm run verify:claude-rewind && npm run verify:layout-tree && npm run verify:fork-session && npm run verify:composer-link-chip && npm run verify:run-group && npm run verify:subagent-workstream && npm run verify:prompt-history && npm run verify:claude-turn-latency && npm run verify:claude-interrupt-stop && npm run verify:subagent-panel && npm run verify:stream-perf && npm run verify:design-writeback",
+    "test": "npm run validate:terminal && npm run check:icons && npm run verify:review-diff-panel && npm run verify:workstream-stages && npm run verify:workstream-preamble-order && npm run verify:picker-menu-side && npm run verify:environment-hub-anchor && npm run verify:git-branch-pill && npm run verify:side-chat-drop && npm run verify:opencode-sdk-adapter && npm run verify:pi-sdk-adapter && npm run verify:codex-compaction && npm run verify:spreadsheet-preview && npm run verify:claude-rewind && npm run verify:layout-tree && npm run verify:fork-session && npm run verify:composer-link-chip && npm run verify:run-group && npm run verify:subagent-workstream && npm run verify:prompt-history && npm run verify:claude-turn-latency && npm run verify:claude-interrupt-stop && npm run verify:subagent-panel && npm run verify:stream-perf && npm run verify:design-mode",
     "verify:subagent-panel": "node scripts/verify-subagent-panel.mjs",
-    "verify:design-writeback": "node scripts/verify-design-writeback.mjs",
+    "verify:design-mode": "node scripts/verify-design-mode.mjs",
     "verify:design-e2e": "node scripts/verify-design-e2e.mjs",
     "verify:layout-tree": "node scripts/verify-layout-tree.mjs",
     "verify:subagent-workstream": "node scripts/verify-subagent-workstream.mjs",
@@ -49,7 +49,6 @@
   "dependencies": {
     "@anthropic-ai/claude-agent-sdk": "^0.3.197",
     "@anthropic-ai/sdk": "^0.98.0",
-    "@babel/parser": "^7.29.7",
     "@base-ui-components/react": "^1.0.0-rc.0",
     "@codemirror/commands": "^6.10.3",
     "@codemirror/lang-markdown": "^6.5.0",
@@ -81,7 +80,6 @@
     "exceljs": "^4.4.0",
     "fflate": "^0.8.3",
     "katex": "^0.16.28",
-    "magic-string": "^0.30.21",
     "motion": "^12.27.5",
     "node-pty": "^1.1.0",
     "react": "^19.0.0",
